% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEWorkshopEuroBioc2025-pkg.R
\docType{package}
\name{iSEEWorkshopEuroBioc2025-pkg}
\alias{iSEEWorkshopEuroBioc2025}
\alias{iSEEWorkshopEuroBioc2025-package}
\alias{iSEEWorkshopEuroBioc2025-pkg}
\title{iSEEWorkshopEuroBioc2025: Learning to use the package iSEE}
\description{
A two-hours session to learn how to use in a reasonable depth the iSEE
package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iSEE.github.io/iSEEWorkshopEuroBioc2025/}
  \item Report bugs at \url{https://github.com/iSEE/iSEEWorkshopEuroBioc2025/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

Authors:
\itemize{
  \item Kevin Rue-Albrecht \email{kevinrue67@gmail.com} (\href{https://orcid.org/0000-0003-3899-3872}{ORCID})
  \item Charlotte Soneson \email{charlottesoneson@gmail.com} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID})
  \item Najla Abassi \email{abassi.nejla96@gmail.com} (\href{https://orcid.org/0000-0001-8357-0938}{ORCID})
}

}
\keyword{internal}
